# perf: only install canister once when running benchmarks